feat: move code to rustdesk, flatpak is not compatible with password
