+++ internal/agent/rustdesk/rustdesk_darwin.go
@@ -2,20 +2,6 @@
 
 package rustdesk
 
-type RustDeskConfig struct {
-	Binary            string
-	LaunchArgs        []string
-	GetIDArgs         []string
-	ConfigFile        string
-	Password          string
-	UseDirectIPAccess bool
-	Whitelist         string
-}
-
-func New() *RustDeskConfig {
-	return &RustDeskConfig{}
-}
-
 func (cfg *RustDeskConfig) Configure(config []byte) error {
 	return nil
 }
@@ -25,10 +11,6 @@ func (cfg *RustDeskConfig) GetInstallationInfo() error {
 }
 
 func (cfg *RustDeskConfig) LaunchRustDesk() error {
-	return nil
-}
-
-func (cfg *RustDeskConfig) SetRustDeskPassword(config []byte) error {
 	return nil
 }
