+++ internal/agent/rustdesk/rustdesk_linux.go
@@ -19,28 +19,6 @@ import (
 	"github.com/shirou/gopsutil/v3/process"
 )
 
-type RustDeskConfig struct {
-	User              *RustDeskUser
-	Binary            string
-	LaunchArgs        []string
-	GetIDArgs         []string
-	ConfigFile        string
-	IsFlatpak         bool
-	UseDirectIPAccess bool
-	Whitelist         string
-}
-
-type RustDeskUser struct {
-	Username string
-	Uid      int
-	Gid      int
-	Home     string
-}
-
-func New() *RustDeskConfig {
-	return &RustDeskConfig{}
-}
-
 func (cfg *RustDeskConfig) GetInstallationInfo() error {
 	rdUser, err := getRustDeskUserInfo()
 	if err != nil {
@@ -101,8 +79,7 @@ func (cfg *RustDeskConfig) Configure(config []byte) error {
 		configPath = filepath.Join(rootConfigPath, "app", "com.rustdesk.RustDesk", "config", "rustdesk")
 		configFile = filepath.Join(configPath, "RustDesk2.toml")
 	} else {
-		rootConfigPath = "/root/.config/rustdesk"
-		// rootConfigPath = filepath.Join(cfg.User.Home, ".config", "rustdesk")
+		rootConfigPath = filepath.Join(cfg.User.Home, ".config", "rustdesk")
 		configPath = rootConfigPath
 		configFile = filepath.Join(configPath, "RustDesk2.toml")
 	}
@@ -138,21 +115,17 @@ func (cfg *RustDeskConfig) Configure(config []byte) error {
 		}
 	}
 
-	if err := os.MkdirAll(configPath, 0755); err != nil {
-		log.Printf("[ERROR]: could not create directory file for RustDesk configuration, reason: %v", err)
-		return err
-	}
-
 	if cfg.IsFlatpak {
+		if err := os.MkdirAll(configPath, 0755); err != nil {
+			log.Printf("[ERROR]: could not create directory file for RustDesk configuration, reason: %v", err)
+			return err
+		}
+
 		if err := ChownRecursively(rootConfigPath, cfg.User.Uid, cfg.User.Gid); err != nil {
 			log.Printf("[ERROR]: could not chown directory file for RustDesk configuration, reason: %v", err)
 			return err
 		}
-	} else {
-		if err := ChownRecursively(rootConfigPath, 0, 0); err != nil {
-			log.Printf("[ERROR]: could not chown directory file for RustDesk configuration, reason: %v", err)
-			return err
-		}
+
 	}
 
 	if err := os.WriteFile(configFile, rdTOML, 0600); err != nil {
@@ -167,11 +140,13 @@ func (cfg *RustDeskConfig) Configure(config []byte) error {
 		}
 	}
 
-	// Restart the RustDesk service to apply the new configuration
-	command := "/usr/bin/systemctl restart rustdesk"
-	cmd := exec.Command("bash", "-c", command)
-	if err := cmd.Run(); err != nil {
-		return err
+	// Restart the RustDesk service to apply the new configuration if not flatpak
+	if !cfg.IsFlatpak {
+		command := "/usr/bin/systemctl restart rustdesk"
+		cmd := exec.Command("bash", "-c", command)
+		if err := cmd.Run(); err != nil {
+			return err
+		}
 	}
 
 	return nil
@@ -179,63 +154,6 @@ func (cfg *RustDeskConfig) Configure(config []byte) error {
 
 func (cfg *RustDeskConfig) LaunchRustDesk() error {
 	return runtime.RunAsUserInBackground(cfg.User.Username, cfg.Binary, cfg.LaunchArgs, true)
-}
-
-func (cfg *RustDeskConfig) SetRustDeskPassword(config []byte) error {
-
-	// The --password command requires root privileges which is not
-	// possible using Flatpak so we've to do a workaround
-	// creating an empty RustDesk.toml file with the password
-	// encrypted
-
-	// Unmarshal configuration data
-	var rdConfig nats.RustDesk
-	if err := json.Unmarshal(config, &rdConfig); err != nil {
-		log.Println("[ERROR]: could not unmarshall RustDesk configuration")
-		return err
-	}
-
-	// Configuration file location
-	configFile := ""
-	rootConfigPath := ""
-	configPath := ""
-
-	if cfg.IsFlatpak {
-		rootConfigPath = filepath.Join(cfg.User.Home, ".var")
-		configPath = filepath.Join(rootConfigPath, "app", "com.rustdesk.RustDesk", "config", "rustdesk")
-		configFile = filepath.Join(configPath, "RustDesk.toml")
-	} else {
-		rootConfigPath = filepath.Join(cfg.User.Home, ".config", "rustdesk")
-		configPath = rootConfigPath
-		configFile = filepath.Join(configPath, "RustDesk.toml")
-	}
-
-	currentConfig, err := os.ReadFile(configFile)
-	if err != nil {
-		log.Printf("[ERROR]: could not read RustDesk.toml file, reason: %v", err)
-		return err
-	}
-
-	tomlConfig := RustDeskPassword{}
-	if err := toml.Unmarshal(currentConfig, &tomlConfig); err != nil {
-		log.Printf("[ERROR]: could not unmarshal RustDesk.toml file, reason: %v", err)
-		return err
-	}
-
-	tomlConfig.Password = rdConfig.PermanentPassword
-
-	data, err := toml.Marshal(tomlConfig)
-	if err != nil {
-		log.Printf("[ERROR]: could not marshal new configuration file, reason: %v", err)
-		return err
-	}
-
-	if err := os.WriteFile(configFile, data, 0600); err != nil {
-		log.Printf("[ERROR]: could not write configuration file with new password, reason: %v", err)
-		return err
-	}
-
-	return nil
 }
 
 func (cfg *RustDeskConfig) GetRustDeskID() (string, error) {
